+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Features:
 
   - Add a 'warnings' variable/method to ConfigValidator.
   - Support simple parameter deprecations in ConfigValidator.
+  - Log schema warnings encountered in ConfigLoader.
 
 ## 0.3.0 (2014-05-13)